% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkme_data-package.R
\docType{package}
\name{benchmarkmeData-package}
\alias{benchmarkmeData-package}
\title{benchmarkmeData: Data Set for the 'benchmarkme' Package}
\description{
Crowd sourced benchmarks from running the 'benchmarkme' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/csgillespie/benchmarkme-data}
  \item Report bugs at \url{https://github.com/csgillespie/benchmarkme-data/issues}
}

}
\author{
\strong{Maintainer}: Colin Gillespie \email{csgillespie@gmail.com} (\href{https://orcid.org/0000-0003-1787-0275}{ORCID})

}
\keyword{internal}
